requirements.txt
streamlit
pandas
pyarrow
plotly
numpy
Pillow
scikit-learn